feat: 채팅방 입장 API 변경
