+++ src/components/class/CreateChat.jsx
@@ -1,0 +1,5 @@
+const CreateChat = (props) => {
+  return <>CC</>;
+};
+
+export default CreateChat;
